requirements.txt
fastapi
unicorn
pymongo
motor  
python-jose  
passlib  
jinja2
fastapi-mail